Fix build caused by rational-solver2.h casting to (long long int) instead of (int64_t) on line 443

+++ linbox/linbox/algorithms/rational-solver2.h
@@ -440,7 +440,7 @@ namespace LinBox
 			for (p_x = x; p_x != x + n; ++ p_x, ++ p_A) {
 				//mpz_set_d (tmp, *p_A);
 				//mpz_addmul_si (*p_y, *p_x, (int)(*p_A));
-				tmp = *p_x  * (long long int)(*p_A);
+				tmp = *p_x  * (int64_t)(*p_A);
 				integer::addin (*p_y, tmp);
 			}
 		}
